multi: fix lint errors across packages
Address golangci-lint findings: remove unnecessary int32 conversion
in ListRounds, use errors.Is for io.EOF check in rounds watch,
add nolint:funlen to processOutbox dispatch loop, wrap long comment
in actormsg, lowercase error string in cmd_board, remove leading
newline in cmd_wallet, and drop stale nolint:funlen from
RunUntilShutdown.

+++ darepod/rpc_server.go
@@ -1092,7 +1092,7 @@ func (r *RPCServer) ListRounds(ctx context.Context,
 	}
 
 	// Query persisted rounds from SQL with cursor pagination.
-	pageSize := int32(req.PageSize)
+	pageSize := req.PageSize
 	if pageSize <= 0 {
 		pageSize = defaultListRoundsPageSize
 	}
